requirements.txt
numpy>=1.24
pandas>=1.5
PyYAML>=6.0
scikit-learn>=1.3
optuna>=3.6
# Deep‑Learning‑Stack
torch>=2.2        # CUDA‑Build (cu118/cu121) installieren!
pytorch-lightning>=2.2
rtdl==0.14.0.dev5   
transformers>=4.43
accelerate>=0.28
datasets>=2.19
bitsandbytes>=0.43
peft>=0.10
# Gradient‑Boosting‑Familie
xgboost>=2.0
lightgbm>=4.3
catboost>=1.2
ngboost>=0.3.12
interpret>=0.4
pytorch-tabnet>=4.1.0
# Utilities
joblib>=1.4
tqdm>=4.66
lightgbm>=4.6
xgboost>=3.0      # für Schritt 2
pytorch-lightning>=2.5
optuna>=4.4       # schon vorhanden
